// src/motion-provider/types.ts
export interface AnimationObjProps {
  [key: string]: unknown;
}
export type AnimationModule = {
  initial: AnimationObjProps;
  animate: AnimationObjProps;
};